[identities] Avoid weird error about not creating SortingHat database
SortingHat had no way of finding out whether you want to reuse its
database (that is, using an existing database) when calling Init.
Now, we have --reuse option. This patch calls sortinghat's Init
with --reuse always, so that it assumes that if the database
exists, it was already initialized by SortingHat. The assumption
was already the same, but now we avoid the weird message.
(It was very strange seeing mordred running with no trouble,
but seeing it print "Error: database could not be created" messages).