[task_panels] Catch name or service not found error when creating .kibana
For creating .kibana in Kibana 6, there is a hardwired host name,
which is supposed to host the Kibiter instance. While we fix this
with conffiguration options, I change the scope of the exception handler
to catch this exception, so that it works if Kibiter is hosted in
a host with some other name.

This is temporary, config options would be very much preferred.